pyproject.toml
[tool.poetry]
name = "prisons"
version = "0.1.0"
description = ""
authors = ["mrmamongo <mrmamongo@engee.cloud>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = {extras = ["mypy"], version = "^2.0.19"}
nats-py = "^2.3.1"
pydantic = "^2.0.3"
pydantic-settings = "^2.0.2"
black = "^23.7.0"
aiosqlite = "^0.19.0"
fastapi = "^0.100.0"
uvicorn = "^0.23.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
